# Faster FMA on Haswell